requirements.txt
Django==5.0.6
djangorestframework==3.15.2
pillow==10.3.0
six==1.16.0
psycopg2-binary==2.9.9
django-storages==1.14.4
boto3==1.34.148
django-cors-headers==4.4.0
